+++ itou/utils/tests.py
@@ -1,22 +1,16 @@
-import copy
 import datetime
 import decimal
 import functools
 import importlib
 import json
-import logging
 import uuid
-from unittest import mock
 
 import faker
-import httpx
 import pytest
-import respx
 from dateutil.relativedelta import relativedelta
 from django.conf import settings
 from django.contrib.auth.models import Permission
 from django.contrib.contenttypes.models import ContentType
-from django.contrib.gis.geos import GEOSGeometry
 from django.contrib.sessions.middleware import SessionMiddleware
 from django.core.exceptions import ValidationError
 from django.core.mail.message import EmailMessage
@@ -43,18 +37,6 @@ from itou.users.enums import KIND_JOB_SEEKER, KIND_PRESCRIBER, KIND_SIAE_STAFF
 from itou.users.factories import JobSeekerFactory, PrescriberFactory, UserFactory
 from itou.users.models import User
 from itou.utils import constants as global_constants, pagination
-from itou.utils.apis import api_entreprise
-from itou.utils.apis.exceptions import GeocodingDataError
-from itou.utils.apis.geocoding import get_geocoding_data
-from itou.utils.apis.pole_emploi import PoleEmploiAPIBadResponse, PoleEmploiApiClient, PoleEmploiAPIException
-from itou.utils.mocks.api_entreprise import ETABLISSEMENT_API_RESULT_MOCK, INSEE_API_RESULT_MOCK
-from itou.utils.mocks.geocoding import BAN_GEOCODING_API_NO_RESULT_MOCK, BAN_GEOCODING_API_RESULT_MOCK
-from itou.utils.mocks.pole_emploi import (
-    API_MAJPASS_RESULT_ERROR,
-    API_MAJPASS_RESULT_OK,
-    API_RECHERCHE_ERROR,
-    API_RECHERCHE_RESULT_KNOWN,
-)
 from itou.utils.models import PkSupportRemark
 from itou.utils.password_validation import CnilCompositionPasswordValidator
 from itou.utils.perms.context_processors import get_current_organization_and_perms
@@ -273,35 +255,6 @@ class ContextProcessorsGetCurrentOrganizationAndPermsTest(TestCase):
                     "account_current_institution_id": institution2.pk,
                 },
             }
-
-
-class UtilsGeocodingTest(TestCase):
-    @mock.patch("itou.utils.apis.geocoding.call_ban_geocoding_api", return_value=BAN_GEOCODING_API_RESULT_MOCK)
-    def test_get_geocoding_data(self, mock_call_ban_geocoding_api):
-        geocoding_data = mock_call_ban_geocoding_api()
-        result = get_geocoding_data(geocoding_data)
-        # Expected data comes from BAN_GEOCODING_API_RESULT_MOCK.
-        expected = {
-            "score": 0.587663373207207,
-            "address_line_1": "10 Pl des Cinq Martyrs du Lycee Buffon",
-            "number": "10",
-            "lane": "Pl des Cinq Martyrs du Lycee Buffon",
-            "address": "10 Pl des Cinq Martyrs du Lycee Buffon",
-            "post_code": "75015",
-            "insee_code": "75115",
-            "city": "Paris",
-            "longitude": 2.316754,
-            "latitude": 48.838411,
-            "coords": GEOSGeometry("POINT(2.316754 48.838411)"),
-        }
-        self.assertEqual(result, expected)
-
-    @mock.patch("itou.utils.apis.geocoding.call_ban_geocoding_api", return_value=BAN_GEOCODING_API_NO_RESULT_MOCK)
-    def test_get_geocoding_data_error(self, mock_call_ban_geocoding_api):
-        geocoding_data = mock_call_ban_geocoding_api()
-
-        with self.assertRaises(GeocodingDataError):
-            get_geocoding_data(geocoding_data)
 
 
 class UtilsValidatorsTest(TestCase):
@@ -839,316 +792,6 @@ class CnilCompositionPasswordValidatorTest(SimpleTestCase):
 
     def test_help_text(self):
         self.assertEqual(CnilCompositionPasswordValidator().get_help_text(), CnilCompositionPasswordValidator.HELP_MSG)
-
-
-@override_settings(
-    API_INSEE_BASE_URL="https://fake.insee.url", API_INSEE_CONSUMER_KEY="foo", API_INSEE_CONSUMER_SECRET="bar"
-)
-class INSEEApiTest(SimpleTestCase):
-    @respx.mock
-    def test_access_token(self):
-        endpoint = respx.post(f"{settings.API_INSEE_BASE_URL}/token").respond(200, json=INSEE_API_RESULT_MOCK)
-
-        access_token = api_entreprise.get_access_token()
-
-        self.assertTrue(endpoint.called)
-        self.assertIn(b"grant_type=client_credentials", endpoint.calls.last.request.content)
-        self.assertTrue(endpoint.calls.last.request.headers["Authorization"].startswith("Basic "))
-        self.assertEqual(access_token, INSEE_API_RESULT_MOCK["access_token"])
-
-    @respx.mock
-    def test_access_token_with_http_error(self):
-        respx.post(f"{settings.API_INSEE_BASE_URL}/token").respond(400)
-
-        with self.assertLogs(api_entreprise.logger, logging.ERROR) as cm:
-            access_token = api_entreprise.get_access_token()
-
-        self.assertIsNone(access_token)
-        self.assertIn("Failed to retrieve an access token", cm.records[0].message)
-        self.assertIs(cm.records[0].exc_info[0], httpx.HTTPStatusError)
-
-    @respx.mock
-    def test_access_token_with_json_error(self):
-        respx.post(f"{settings.API_INSEE_BASE_URL}/token").respond(200, content=b"not-json")
-
-        with self.assertLogs(api_entreprise.logger, logging.ERROR) as cm:
-            access_token = api_entreprise.get_access_token()
-
-        self.assertIsNone(access_token)
-        self.assertIn("Failed to retrieve an access token", cm.records[0].message)
-        self.assertIs(cm.records[0].exc_info[0], json.JSONDecodeError)
-
-
-@override_settings(
-    API_INSEE_BASE_URL="https://fake.insee.url",
-    API_ENTREPRISE_BASE_URL="https://api.entreprise.fake.com",
-    API_INSEE_CONSUMER_KEY="foo",
-    API_INSEE_CONSUMER_SECRET="bar",
-)
-class ApiEntrepriseTest(SimpleTestCase):
-    def setUp(self):
-        super().setUp()
-
-        self.token_endpoint = respx.post(f"{settings.API_INSEE_BASE_URL}/token").respond(
-            200,
-            json=INSEE_API_RESULT_MOCK,
-        )
-
-        self.siret_endpoint = respx.get(f"{settings.API_ENTREPRISE_BASE_URL}/siret/26570134200148")
-
-    @respx.mock
-    def test_etablissement_get_or_error(self):
-        self.siret_endpoint.respond(200, json=ETABLISSEMENT_API_RESULT_MOCK)
-
-        etablissement, error = api_entreprise.etablissement_get_or_error("26570134200148")
-
-        self.assertIsNone(error)
-        self.assertEqual(etablissement.name, "CENTRE COMMUNAL D'ACTION SOCIALE")
-        self.assertEqual(etablissement.address_line_1, "22 RUE DU WAD BILLY")
-        self.assertEqual(etablissement.address_line_2, "22-24")
-        self.assertEqual(etablissement.post_code, "57000")
-        self.assertEqual(etablissement.city, "METZ")
-        self.assertEqual(etablissement.department, "57")
-        self.assertFalse(etablissement.is_closed)
-        self.assertTrue(etablissement.is_head_office)
-
-    @respx.mock
-    def test_etablissement_get_or_error_with_closed_status(self):
-        data = copy.deepcopy(ETABLISSEMENT_API_RESULT_MOCK)
-        data["etablissement"]["periodesEtablissement"][0]["etatAdministratifEtablissement"] = "F"
-        self.siret_endpoint.respond(200, json=data)
-
-        etablissement, error = api_entreprise.etablissement_get_or_error("26570134200148")
-
-        self.assertIsNone(error)
-        self.assertTrue(etablissement.is_closed)
-
-    @respx.mock
-    def test_etablissement_get_or_error_without_token(self):
-        self.token_endpoint.respond(404)
-
-        result = api_entreprise.etablissement_get_or_error("whatever")
-
-        self.assertEqual(result, (None, "Problème de connexion à la base Sirene. Essayez ultérieurement."))
-
-    @respx.mock
-    def test_etablissement_get_or_error_with_request_error(self):
-        self.siret_endpoint.mock(side_effect=httpx.RequestError)
-
-        with self.assertLogs(api_entreprise.logger, logging.ERROR) as cm:
-            result = api_entreprise.etablissement_get_or_error("26570134200148")
-
-        self.assertEqual(result, (None, "Problème de connexion à la base Sirene. Essayez ultérieurement."))
-        self.assertTrue(cm.records[0].message.startswith("A request to the INSEE API failed"))
-
-    @respx.mock
-    def test_etablissement_get_or_error_with_other_http_bad_request_error(self):
-        self.siret_endpoint.respond(400)
-
-        result = api_entreprise.etablissement_get_or_error("26570134200148")
-
-        self.assertEqual(result, (None, "Erreur dans le format du SIRET : « 26570134200148 »."))
-
-    @respx.mock
-    def test_etablissement_get_or_error_with_other_http_forbidden_error(self):
-        self.siret_endpoint.respond(403)
-
-        result = api_entreprise.etablissement_get_or_error("26570134200148")
-
-        self.assertEqual(result, (None, "Cette entreprise a exercé son droit d'opposition auprès de l'INSEE."))
-
-    @respx.mock
-    def test_etablissement_get_or_error_with_other_http_not_found_error(self):
-        self.siret_endpoint.respond(404)
-
-        result = api_entreprise.etablissement_get_or_error("26570134200148")
-
-        self.assertEqual(result, (None, "SIRET « 26570134200148 » non reconnu."))
-
-    @respx.mock
-    def test_etablissement_get_or_error_with_http_error(self):
-        self.siret_endpoint.respond(401)
-
-        with self.assertLogs(api_entreprise.logger, logging.ERROR) as cm:
-            result = api_entreprise.etablissement_get_or_error("26570134200148")
-
-        self.assertEqual(result, (None, "Problème de connexion à la base Sirene. Essayez ultérieurement."))
-        self.assertTrue(cm.records[0].message.startswith("Error while fetching"))
-
-    @respx.mock
-    def test_etablissement_get_or_error_when_content_is_not_json(self):
-        self.siret_endpoint.respond(200, content=b"not-json")
-
-        with self.assertLogs(api_entreprise.logger, logging.ERROR) as cm:
-            result = api_entreprise.etablissement_get_or_error("26570134200148")
-
-        self.assertEqual(result, (None, "Le format de la réponse API Entreprise est non valide."))
-        self.assertTrue(cm.records[0].message.startswith("Invalid format of response from API Entreprise"))
-        self.assertIs(cm.records[0].exc_info[0], json.JSONDecodeError)
-
-    @respx.mock
-    def test_etablissement_get_or_error_when_content_is_missing_data(self):
-        self.siret_endpoint.respond(200, json={})
-
-        with self.assertLogs(api_entreprise.logger, logging.ERROR) as cm:
-            result = api_entreprise.etablissement_get_or_error("26570134200148")
-
-        self.assertEqual(result, (None, "Le format de la réponse API Entreprise est non valide."))
-        self.assertTrue(cm.records[0].message.startswith("Invalid format of response from API Entreprise"))
-        self.assertIs(cm.records[0].exc_info[0], KeyError)
-
-    @respx.mock
-    def test_etablissement_get_or_error_when_content_is_missing_historical_data(self):
-        data = copy.deepcopy(ETABLISSEMENT_API_RESULT_MOCK)
-        data["etablissement"]["periodesEtablissement"] = []
-        self.siret_endpoint.respond(200, json=data)
-
-        with self.assertLogs(api_entreprise.logger, logging.ERROR) as cm:
-            result = api_entreprise.etablissement_get_or_error("26570134200148")
-
-        self.assertEqual(result, (None, "Le format de la réponse API Entreprise est non valide."))
-        self.assertTrue(cm.records[0].message.startswith("Invalid format of response from API Entreprise"))
-        self.assertIs(cm.records[0].exc_info[0], IndexError)
-
-    @respx.mock
-    def test_etablissement_get_or_error_with_missing_address_number(self):
-        data = copy.deepcopy(ETABLISSEMENT_API_RESULT_MOCK)
-        data["etablissement"]["adresseEtablissement"]["numeroVoieEtablissement"] = None
-        self.siret_endpoint.respond(200, json=data)
-
-        etablissement, error = api_entreprise.etablissement_get_or_error("26570134200148")
-
-        self.assertIsNone(error)
-        self.assertEqual(etablissement.address_line_1, "RUE DU WAD BILLY")
-
-    @respx.mock
-    def test_etablissement_get_or_error_with_empty_address(self):
-        data = copy.deepcopy(ETABLISSEMENT_API_RESULT_MOCK)
-        data["etablissement"]["adresseEtablissement"] = {
-            "complementAdresseEtablissement": None,
-            "numeroVoieEtablissement": None,
-            "typeVoieEtablissement": None,
-            "libelleVoieEtablissement": None,
-            "codePostalEtablissement": None,
-            "libelleCommuneEtablissement": None,
-            "codeCommuneEtablissement": None,
-        }
-        self.siret_endpoint.respond(200, json=data)
-
-        etablissement, error = api_entreprise.etablissement_get_or_error("26570134200148")
-
-        self.assertIsNone(error)
-        self.assertEqual(etablissement.address_line_1, None)
-        self.assertEqual(etablissement.address_line_2, None)
-        self.assertEqual(etablissement.post_code, None)
-        self.assertEqual(etablissement.city, None)
-        self.assertEqual(etablissement.department, None)
-
-
-class PoleEmploiAPIClientTest(TestCase):
-    def setUp(self) -> None:
-        self.api_client = PoleEmploiApiClient(
-            "https://some.auth.domain", "https://some-authentication-domain.fr", "foobar", "pe-secret"
-        )
-        respx.post(self.api_client.token_url).respond(
-            200, json={"token_type": "foo", "access_token": "batman", "expires_in": 3600}
-        )
-
-    @respx.mock
-    def test_get_token_nominal(self):
-        now = timezone.now()
-        self.api_client._refresh_token(at=now)
-        self.assertEqual(self.api_client.token, "foo batman")
-        self.assertEqual(self.api_client.expires_at, now + datetime.timedelta(seconds=3600))
-
-    @respx.mock
-    def test_get_token_fails(self):
-        job_seeker = JobSeekerFactory()
-        respx.post(self.api_client.token_url).mock(side_effect=httpx.ConnectTimeout)
-        with self.assertRaises(PoleEmploiAPIException) as ctx:
-            self.api_client.recherche_individu_certifie(
-                job_seeker.first_name, job_seeker.last_name, job_seeker.birthdate, job_seeker.nir
-            )
-        self.assertEqual(ctx.exception.error_code, "http_error")
-
-    @respx.mock
-    def test_recherche_individu_certifie_api_nominal(self):
-        job_seeker = JobSeekerFactory()
-        respx.post(self.api_client.recherche_individu_url).respond(200, json=API_RECHERCHE_RESULT_KNOWN)
-        id_national = self.api_client.recherche_individu_certifie(
-            job_seeker.first_name, job_seeker.last_name, job_seeker.birthdate, job_seeker.nir
-        )
-        self.assertEqual(id_national, "ruLuawDxNzERAFwxw6Na4V8A8UCXg6vXM_WKkx5j8UQ")
-
-        # now with weird payloads
-        job_seeker.first_name = "marié%{-christine}  aéïèêë " + "a" * 50
-        job_seeker.last_name = "gh'îkñ Bind-n'qici " + "b" * 50
-        id_national = self.api_client.recherche_individu_certifie(
-            job_seeker.first_name, job_seeker.last_name, job_seeker.birthdate, job_seeker.nir
-        )
-        payload = json.loads(respx.calls.last.request.content)
-        self.assertEqual(payload["nomNaissance"], "GH'IKN BIND-N'QICI BBBBBB")  # 25 chars
-        self.assertEqual(payload["prenom"], "MARIE-CHRISTI")  # 13 chars
-
-    @respx.mock
-    def test_recherche_individu_certifie_individual_api_errors(self):
-        job_seeker = JobSeekerFactory()
-        respx.post(self.api_client.recherche_individu_url).respond(200, json=API_RECHERCHE_ERROR)
-        with self.assertRaises(PoleEmploiAPIBadResponse) as ctx:
-            self.api_client.recherche_individu_certifie(
-                job_seeker.first_name, job_seeker.last_name, job_seeker.birthdate, job_seeker.nir
-            )
-        self.assertEqual(ctx.exception.response_code, "R010")
-
-    @respx.mock
-    def test_recherche_individu_certifie_retryable_errors(self):
-        job_seeker = JobSeekerFactory()
-
-        respx.post(self.api_client.recherche_individu_url).respond(401, json="")
-        with self.assertRaises(PoleEmploiAPIException) as ctx:
-            self.api_client.recherche_individu_certifie(
-                job_seeker.first_name, job_seeker.last_name, job_seeker.birthdate, job_seeker.nir
-            )
-        self.assertEqual(ctx.exception.error_code, 401)
-
-        job_seeker = JobSeekerFactory()
-        respx.post(self.api_client.recherche_individu_url).mock(side_effect=httpx.ConnectTimeout)
-        with self.assertRaises(PoleEmploiAPIException) as ctx:
-            self.api_client.recherche_individu_certifie(
-                job_seeker.first_name, job_seeker.last_name, job_seeker.birthdate, job_seeker.nir
-            )
-        self.assertEqual(ctx.exception.error_code, "http_error")
-
-    @respx.mock
-    def test_mise_a_jour_pass_iae_success_with_approval_accepted(self):
-        """
-        Nominal scenario: an approval is **accepted**
-        HTTP 200 + codeSortie = S001 is the only way mise_a_jour_pass_iae does not raise.
-        """
-        job_application = JobApplicationFactory(with_approval=True)
-        respx.post(self.api_client.mise_a_jour_url).respond(200, json=API_MAJPASS_RESULT_OK)
-        # we really don't care about the arguments there
-        self.api_client.mise_a_jour_pass_iae(job_application.approval, "foo", "bar", 42, "DEAD")
-
-    @respx.mock
-    def test_mise_a_jour_pass_iae_failure(self):
-        job_application = JobApplicationFactory(with_approval=True)
-        # non-S001 codeSortie
-        respx.post(self.api_client.mise_a_jour_url).respond(200, json=API_MAJPASS_RESULT_ERROR)
-        with self.assertRaises(PoleEmploiAPIBadResponse):
-            self.api_client.mise_a_jour_pass_iae(job_application.approval, "foo", "bar", 42, "DEAD")
-
-        # timeout
-        respx.post(self.api_client.mise_a_jour_url).mock(side_effect=httpx.ConnectTimeout)
-        with self.assertRaises(PoleEmploiAPIException) as ctx:
-            self.api_client.mise_a_jour_pass_iae(job_application.approval, "foo", "bar", 42, "DEAD")
-        self.assertEqual(ctx.exception.error_code, "http_error")
-
-        # auth failed
-        respx.post(self.api_client.mise_a_jour_url).respond(401, json={})
-        with self.assertRaises(PoleEmploiAPIException):
-            self.api_client.mise_a_jour_pass_iae(job_application.approval, "foo", "bar", 42, "DEAD")
-        self.assertEqual(ctx.exception.error_code, "http_error")
 
 
 class UtilsEmailsSplitRecipientTest(TestCase):
